crud in memory implementation

+++ src/main/java/ru/javawebinar/topjava/repository/mock/InMemoryMealRepositoryImpl.java
@@ -4,9 +4,9 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Repository;
 import ru.javawebinar.topjava.model.Meal;
-import ru.javawebinar.topjava.model.User;
 import ru.javawebinar.topjava.repository.MealRepository;
 import ru.javawebinar.topjava.util.MealsUtil;
+import ru.javawebinar.topjava.util.exception.NotFoundException;
 
 import java.util.*;
 import java.util.concurrent.ConcurrentHashMap;
@@ -18,7 +18,6 @@ public class InMemoryMealRepositoryImpl implements MealRepository {
 
     protected final Logger log = LoggerFactory.getLogger(getClass());
     private Map<Integer, Meal> repository = new ConcurrentHashMap<>();
-    private Map<Integer, User> users = new ConcurrentHashMap<>();
     private AtomicInteger counter = new AtomicInteger(0);
 
     {
@@ -38,25 +37,37 @@ public class InMemoryMealRepositoryImpl implements MealRepository {
     }
 
     @Override
-    public boolean delete(int id) {
+    public boolean delete(int id, Integer userId) throws NotFoundException{
         log.info("delete meal {}", id);
-        return repository.remove(id, repository.get(id));
+        Meal meal = repository.get(id);
+        if (Objects.equals(meal.getUserId(), userId))
+            return repository.remove(id, repository.get(id));
+        else
+            throw new NotFoundException("Cant delete foreign meal");
     }
 
     @Override
-    public Meal get(int id) {
+    public Meal get(int id, Integer userId) {
         log.info("get meal {}", id);
-        return repository.get(id);
+        Meal meal = repository.get(id);
+        if (Objects.equals(meal.getUserId(), userId))
+            return meal;
+        else
+            return null;
     }
 
     @Override
-    public Collection<Meal> getAll() {
+    public Collection<Meal> getAll(Integer userId) {
         log.info("get all meals");
         List<Meal> meals =  repository.values()
                 .stream()
                 .sorted(Comparator.comparing(Meal::getDateTime))
+                .filter(meal -> Objects.equals(meal.getUserId(), userId))
                 .collect(Collectors.toList());
         Collections.reverse(meals);
-        return meals;
+        if (meals.isEmpty())
+            return null;
+        else
+            return meals;
     }
 }
